Attempt - 1: Decleared variable sum and initialized it to zero.
Attempt - 2: introduced a for loop statement: for (let j = 0; j < arrayOfNumbers.length; j++) where a variable, j, is created and zero assigned to it to initialize it. Stated a condition and increment by 1.

Attempt - 3: Introduced a statement, sum += arrayOfNumbers[j], which adds the value of the variable,  arrayOfNumbers[j], to the variable, sum, and assigns the result to the variable, sum.

Attempt - 4: Created a variable countEven and initialized it to zero.

Attempt - 5: Introduced a for ... if loops statements. In the 'for' statement, created a variable, j and initialized it to zero, put condition and set j increment by 1. In the 'if' section of the loops, put a condition that checks if the integer remainder of dividing the number with index j in array of numbers with 2 is equal to zero.

Attempt - 6: increment the variable, countEven by 1.

Attempt - 7: Adjusted the return statement for the right variable.

Attempt - 8: Removed the comment(s) from the line just after the function as directed to make the statement executable.